Day 12 completed.
First time trying to implement a pathfinding algorythm.
Ended up merging a dictionary that had the nodes connections, with a function that would take a tuple from that dictionary and give back a node number.
That way the input would match the one implementation I found here https://stackabuse.com/courses/graphs-in-python-theory-and-implementation/lessons/dijkstras-algorythm
For part 2, I was thinking of just running the algorythm from all the point a to the end, but that would take a long time. So instead , I added a fake node with a weight of 0 to all the a points and made it start there. That gave me the same result, and I only needed to run it once.